nav drawer updates for total bunk and total attend

+++ app/src/main/java/farees/hussain/bunkmanager/db/database/SubjectDao.kt
@@ -17,8 +17,8 @@ interface SubjectDao {
     fun observeAllSubjectItems(): LiveData<List<Subject>>
 
     @Query("select SUM(classesMustAttend) from subject_items")
-    fun observeTotalMustAttendClasses():LiveData<Float>
+    fun observeTotalMustAttendClasses():LiveData<Int>
 
     @Query("select SUM(classesCanBeBunked) from subject_items")
-    fun observeTotalCanBeBunkedClasses():LiveData<Float>
+    fun observeTotalCanBeBunkedClasses():LiveData<Int>
 }
